from Javier Amor Garc��a: """My first contribution is a patch that, following the TODO, "add a modified flag in OutlinerDocument? that tiggers a "Save Changes" message box when quitting or opening a new doc".
One caveat about this patch; when the user starts the outliner or  makes
a new doc, the outliner automatically adds a new element. This behavior
may be  troublesome with the new feature, because just starting the
thing or making the new thing changes the doc. I haven't made a
workaround because it made the code uglier.

The solution may be a change in this behavior or the mentioned
workaround."""