Patch from Javier Amor Garc��a <javier.amor.garcia@gmail.com>:
Here is a patch that add a "uri" attribute to the OutlinerDocument
class and that modifies the outliner_file_save function to use it.
This results in a improved behaviour to the "save" feature that only
prompts the user for a file's name if needed; before it behaved equal
than the "save as" feature.
The "uri" attribute should be too used by future features (like the
future properties dialog or the "load last document at the startup"
feature).